requirements.txt
torch==2.3.1
pillow==10.3.0
pix2tex==0.1.2
